cli: use single config civ.config.jsonc (+ civ.config.schema.json); simplify config lookup and flag text; add tools/graph barrel; fix command imports; update README/FEATURES/AGENTS; rename AGENTS.md
