+++ packages/cli/src/tools/crawler/index.ts
@@ -1,7 +1,6 @@
 export * from './types';
 export { buildIndexFromXml, findXmlFiles } from './xml-indexer';
 export { crawl } from './crawler';
-export { graphToDot, graphToJson } from '../graph/export';
 export { parseSeed } from './seed';
 export * from './constants';
